Reorganise ReactJS
Consolidate components, containers, reducers and actions under module
