+++ web/reactjs/src/modules/actions/fetchAuthUser.js
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { connect } from 'react-redux';
-import { Auth } from '../components';
 import axios from 'axios';
 
 const fetchAuthUser = async (userName, password) => {
@@ -21,22 +19,4 @@ const fetchAuthUser = async (userName, password) => {
     return resp.data;
 };
 
-const mapStateToProps = (store) => {
-    return {
-        user: store.authReducer.user
-    };
-};
-
-const mapDispatchToProps = (dispatch) => {
-    return {
-        authenticate: async (userName, password) => dispatch({
-            type: 'AUTH_USER_FULFILLED', 
-            payload: await fetchAuthUser(userName, password),
-        })
-    };
-};
-
-export default connect(
-    mapStateToProps,
-    mapDispatchToProps
-)(Auth);
+export default fetchAuthUser;
