+++ web/reactjs/src/modules/components/Dashboard/index.js
