+++ web/reactjs/src/modules/components/MainLayout/Footer/index.js
