+++ web/reactjs/src/modules/components/MainLayout/Sidebar/Sidebar.js
