+++ web/reactjs/src/modules/components/MainLayout/Sidebar/components/Profile/Profile.js
