+++ web/reactjs/src/modules/components/MainLayout/Sidebar/index.js
