+++ web/reactjs/src/modules/components/MainLayout/index.js
