+++ web/reactjs/src/modules/components/MinimalLayout/MinimalLayout.js
