+++ web/reactjs/src/modules/components/MinimalLayout/Topbar/index.js
