+++ web/reactjs/src/modules/components/MinimalLayout/index.js
