+++ web/reactjs/src/modules/components/index.js
@@ -14,6 +14,6 @@
 
 export { default as Auth } from './Auth';
 export { default as Dashboard } from './Dashboard';
-export { default as NotFound } from './NotFound';
 export { default as MainLayout } from './MainLayout';
 export { default as MinimalLayout } from './MinimalLayout';
+export { default as NotFound } from './NotFound';
