+++ web/reactjs/src/modules/reducers/authReducer.js
