+++ web/reactjs/src/modules/reducers/index.js
